+++ promgen_k8s/__init__.py
@@ -1,8 +1,9 @@
-__all__ = ['cadvisor_job', 'cluster', 'generator', 'nodes_job', 'pods_job', 'prom_dsl', 'service_endpoints_job', 'services_job']
+__all__ = ['cadvisor_job', 'cluster', 'generator', 'ingresses_job', 'nodes_job', 'pods_job', 'prom_dsl', 'service_endpoints_job', 'services_job']
 
 from cadvisor_job import CadvisorJob
 from cluster import Cluster
 from generator import Generator
+from ingresses_job import IngressesJob
 from nodes_job import NodesJob
 from pods_job import PodsJob
 from prom_dsl import *